Pembuatan Seeder untuk Import Data dari CSV dan Perbaikan Migration

+++ database/migrations/2025_07_15_045611_create_ruangs_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('nama2', 250)->default('-');
             $table->integer('id_jenis_makul')->nullable();
             $table->string('id_fakultas', 5)->nullable();
-            $table->tinyInteger('kapasitas')->default(0);
-            $table->tinyInteger('kap_ujian')->default(0);
+            $table->integer('kapasitas')->default(0);
+            $table->integer('kap_ujian')->default(0);
             $table->tinyInteger('status')->default(1)->comment('1: buka 0: tutup 2: hapus');
             $table->string('luas', 5)->default('0')->comment('meter persegi');
             $table->string('kondisi', 50)->nullable();
